split api-hub.js
