+++ skillboss/scripts/api-hub.js
@@ -1,24 +1,25 @@
 #!/usr/bin/env node
 
-const fs = require('fs')
-const path = require('path')
-const { pipeline } = require('stream/promises')
-const { Readable } = require('stream')
-const { fetchWithRetry } = require('./lib/fetch-retry')
-
 /**
  * SkillBoss API Hub - Multi-Provider API Gateway Client
  *
  * Provides unified access to multiple AI/ML providers:
- * - Chat: bedrock, openai, openrouter, vertex, anthropic, minimax, perplexity
+ * - Chat: bedrock, openai, openrouter, vertex, anthropic, minimax, perplexity, huggingface
  * - TTS: elevenlabs, minimax, openai, replicate
- * - Image: vertex/gemini-3-pro-image-preview, replicate/flux, fal/img2img
+ * - Image: vertex/gemini-3-pro-image-preview, replicate/flux, fal/img2img, huggingface
  * - Upscale: fal/upscale (creative-upscaler)
  * - Search: scrapingdog, perplexity, firecrawl, linkup (structured search + fetch)
- * - Video: minimax
+ * - Video: minimax, huggingface
+ * - STT: openai/whisper-1, huggingface (speech-to-text)
+ * - Embedding: huggingface
  * - Document: reducto (parse, extract, split, edit)
  * - SMS/Verify: prelude (OTP send/check, notify)
  * - Email: aws/ses
+ *
+ * HuggingFace Dynamic Routing:
+ * - Any model on huggingface.co works as "huggingface/{org}/{model}" without pre-registration
+ * - Chat: node api-hub.js chat --model "huggingface/meta-llama/Llama-3-8B-Instruct" --prompt "Hello"
+ * - Other tasks via /run: --inputs '{"task":"embedding"}' or '{"task":"image"}' etc.
  *
  * Usage:
  *   node api-hub.js run --model "vendor/model" --inputs '{"key":"value"}' [--stream] [--output file]
@@ -32,6 +33,7 @@ const { fetchWithRetry } = require('./lib/fetch-retry')
  *   node api-hub.js linkup-search --query "latest AI news" [--output-type searchResults|sourcedAnswer|structured] [--depth standard|deep]
  *   node api-hub.js linkup-fetch --url "https://example.com" [--render-js]
  *   node api-hub.js scrape --model "firecrawl/scrape" --url "https://example.com"
+ *   node api-hub.js stt --file recording.mp3 [--model "openai/whisper-1"] [--prompt "..."] [--language "en"] [--output transcript.txt]
  *   node api-hub.js sms-verify --phone "+1234567890"
  *   node api-hub.js sms-check --phone "+1234567890" --code "123456"
  *   node api-hub.js sms-send --phone "+1234567890" --template-id "your_template_id"
@@ -39,924 +41,21 @@ const { fetchWithRetry } = require('./lib/fetch-retry')
  *   node api-hub.js send-batch --subject "Hello {{name}}" --body "<html>...</html>" --receivers '[...]'
  */
 
-// Load config from config.json (sibling to scripts folder)
-const CONFIG_PATH = path.join(__dirname, '..', 'config.json')
-
-function loadConfig() {
-  try {
-    const configData = fs.readFileSync(CONFIG_PATH, 'utf8')
-    return JSON.parse(configData)
-  } catch (err) {
-    throw new Error(`Failed to load config from ${CONFIG_PATH}: ${err.message}`)
-  }
-}
-
-const config = loadConfig()
-
-// Configuration from config.json
-const API_HUB_API_KEY = config.apiKey
-const API_HUB_BASE_URL = config.baseUrl || 'https://api.heybossai.com/v1'
-
-/**
- * Simple HTTP client for API Hub
- * @param {string} endpoint - API endpoint
- * @param {object} data - Request body
- * @returns {Promise<object>} Response data
- */
-async function apiHubPost(endpoint, data) {
-  if (!API_HUB_API_KEY || API_HUB_API_KEY === 'YOUR_API_KEY_HERE') {
-    throw new Error(
-      'API key not configured. Please update config.json with your API key.',
-    )
-  }
-
-  const response = await fetchWithRetry(`${API_HUB_BASE_URL}${endpoint}`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${API_HUB_API_KEY}`,
-    },
-    body: JSON.stringify(data),
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`API Hub request failed: ${response.status} ${errorText}`)
-  }
-
-  return response.json()
-}
-
-/**
- * Stream response from API Hub (SSE)
- * @param {string} endpoint - API endpoint
- * @param {object} data - Request body
- * @yields {object} Parsed SSE data chunks
- */
-async function* apiHubStream(endpoint, data) {
-  if (!API_HUB_API_KEY || API_HUB_API_KEY === 'YOUR_API_KEY_HERE') {
-    throw new Error(
-      'API key not configured. Please update config.json with your API key.',
-    )
-  }
-
-  const response = await fetchWithRetry(`${API_HUB_BASE_URL}${endpoint}`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${API_HUB_API_KEY}`,
-    },
-    body: JSON.stringify(data),
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`API Hub request failed: ${response.status} ${errorText}`)
-  }
-
-  const reader = response.body.getReader()
-  const decoder = new TextDecoder()
-  let buffer = ''
-
-  while (true) {
-    const { done, value } = await reader.read()
-    if (done) break
-
-    buffer += decoder.decode(value, { stream: true })
-    const lines = buffer.split('\n')
-    buffer = lines.pop() // Keep incomplete line in buffer
-
-    for (const line of lines) {
-      const trimmed = line.trim()
-      if (trimmed.startsWith('data: ')) {
-        const data = trimmed.slice(6)
-        if (data === '[DONE]') return
-        try {
-          yield JSON.parse(data)
-        } catch {
-          // Skip non-JSON data lines
-        }
-      }
-    }
-  }
-}
-
-/**
- * Save binary response to file
- * @param {Response} response - Fetch Response object
- * @param {string} outputPath - File path to save to
- */
-async function saveBinaryResponse(response, outputPath) {
-  const fileStream = fs.createWriteStream(outputPath)
-  await pipeline(Readable.fromWeb(response.body), fileStream)
-}
-
-/**
- * Simple HTTP GET client for API Hub
- * @param {string} endpoint - API endpoint
- * @returns {Promise<object>} Response data
- */
-async function apiHubGet(endpoint) {
-  if (!API_HUB_API_KEY || API_HUB_API_KEY === 'YOUR_API_KEY_HERE') {
-    throw new Error('API key not configured. Please update config.json with your API key.')
-  }
-
-  const response = await fetchWithRetry(`${API_HUB_BASE_URL}${endpoint}`, {
-    method: 'GET',
-    headers: {
-      Authorization: `Bearer ${API_HUB_API_KEY}`,
-    },
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`API Hub request failed: ${response.status} ${errorText}`)
-  }
-
-  return response.json()
-}
-
-/**
- * Make a raw API Hub request that may return binary data
- * @param {string} endpoint - API endpoint
- * @param {object} data - Request body
- * @returns {Promise<Response>} Raw fetch Response
- */
-async function apiHubRaw(endpoint, data) {
-  if (!API_HUB_API_KEY || API_HUB_API_KEY === 'YOUR_API_KEY_HERE') {
-    throw new Error(
-      'API key not configured. Please update config.json with your API key.',
-    )
-  }
-
-  const response = await fetchWithRetry(`${API_HUB_BASE_URL}${endpoint}`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${API_HUB_API_KEY}`,
-    },
-    body: JSON.stringify(data),
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(`API Hub request failed: ${response.status} ${errorText}`)
-  }
-
-  return response
-}
-
-/**
- * Sends a single email using AWS SES via API Hub
- * Sender is automatically determined from user lookup (name@name.skillboss.live)
- * @param {object} params - Email parameters
- * @param {string} params.subject - Email subject line
- * @param {string} params.bodyHtml - Email body in HTML format
- * @param {string[]} params.receivers - Array of recipient email addresses
- * @param {string[]} [params.replyTo] - Reply-to email addresses
- * @param {string} [params.projectId] - Optional project identifier
- * @returns {Promise<object>} Email send status
- */
-async function sendEmail(params) {
-  const data = {
-    title: params.subject,
-    body_html: params.bodyHtml,
-    receivers: params.receivers,
-    project_id: params.projectId,
-  }
-  if (params.replyTo) data.reply_to = params.replyTo
-
-  return apiHubPost('/send-email', data)
-}
-
-/**
- * Sends batch emails with template variables using AWS SES via API Hub
- * Supports {{variable}} template syntax in subject and body
- * Sender is automatically determined from user lookup (name@name.skillboss.live)
- * @param {object} params - Batch email parameters
- * @param {string} params.subject - Email subject line with template variables
- * @param {string} params.bodyHtml - Email body in HTML format with template variables
- * @param {Array<{email: string, variables: object}>} params.receivers - Recipients with template variables
- * @param {string[]} [params.replyTo] - Reply-to email addresses
- * @param {string} [params.projectId] - Optional project identifier
- * @returns {Promise<object>} Batch email results with per-email status
- */
-async function sendBatchEmails(params) {
-  const data = {
-    title: params.subject,
-    body_html: params.bodyHtml,
-    receivers: params.receivers,
-    project_id: params.projectId,
-  }
-  if (params.replyTo) data.reply_to = params.replyTo
-
-  return apiHubPost('/send-emails', data)
-}
-
-// ============================================================================
-// HIGH-LEVEL COMMAND FUNCTIONS
-// ============================================================================
-
-/**
- * Generic run command - mirrors /run endpoint exactly
- * @param {object} params - Run parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {object} params.inputs - Model-specific inputs
- * @param {boolean} [params.stream] - Enable streaming
- * @param {string} [params.output] - Output file path for binary responses
- * @param {boolean} [params.autoFallback] - Enable automatic fallback on errors (default: true)
- * @returns {Promise<object|AsyncGenerator>} Response data or stream
- */
-async function run(params) {
-  const request = {
-    model: params.model,
-    inputs: params.inputs,
-    stream: params.stream || false,
-    auto_fallback: params.autoFallback !== false, // Enable by default
-  }
-
-  if (params.stream) {
-    return apiHubStream('/run', request)
-  }
-
-  if (params.output) {
-    const response = await apiHubRaw('/run', request)
-    const contentType = response.headers.get('content-type') || ''
-
-    if (contentType.includes('audio') || contentType.includes('octet-stream')) {
-      await saveBinaryResponse(response, params.output)
-      return { saved: params.output }
-    }
-    // For JSON responses, check for errors before saving
-    const data = await response.json()
-    if (data.code && data.code >= 400) {
-      throw new Error(data.message || `API error: ${data.code}`)
-    }
-
-    // Check if response contains media URL(s) and download the actual file
-    let mediaUrl = null
-    let mediaType = 'file'
-
-    // Image URL patterns
-    if (
-      Array.isArray(data) &&
-      data.length > 0 &&
-      typeof data[0] === 'string' &&
-      data[0].startsWith('http')
-    ) {
-      // Flux-style response: ["https://..."]
-      mediaUrl = data[0]
-      mediaType = 'image'
-    } else if (data.data && Array.isArray(data.data) && data.data[0]?.url) {
-      // DALL-E style response: {data: [{url: "https://..."}]}
-      mediaUrl = data.data[0].url
-      mediaType = 'image'
-    } else if (
-      data.generated_images &&
-      Array.isArray(data.generated_images) &&
-      data.generated_images[0]
-    ) {
-      // Gemini-style response: {generated_images: ["https://..."]}
-      mediaUrl = data.generated_images[0]
-      mediaType = 'image'
-    } else if (
-      data.image_url &&
-      typeof data.image_url === 'string' &&
-      data.image_url.startsWith('http')
-    ) {
-      // MM-style response: {image_url: "https://..."}
-      mediaUrl = data.image_url
-      mediaType = 'image'
-    }
-    // Audio URL patterns
-    else if (
-      data.audio_url &&
-      typeof data.audio_url === 'string' &&
-      data.audio_url.startsWith('http')
-    ) {
-      // MM TTS response: {audio_url: "https://..."}
-      mediaUrl = data.audio_url
-      mediaType = 'audio'
-    }
-    // Video URL patterns
-    else if (data.video_url) {
-      // Common video response: {video_url: "https://..."}
-      mediaUrl = data.video_url
-      mediaType = 'video'
-    } else if (
-      data.output &&
-      typeof data.output === 'string' &&
-      data.output.startsWith('http')
-    ) {
-      // Replicate-style response: {output: "https://..."}
-      mediaUrl = data.output
-      mediaType = 'video'
-    } else if (
-      data.video &&
-      typeof data.video === 'string' &&
-      data.video.startsWith('http')
-    ) {
-      // Alternative video response: {video: "https://..."}
-      mediaUrl = data.video
-      mediaType = 'video'
-    } else if (data.file_id && data.base_resp?.status_code === 0) {
-      // MiniMax async video - need to poll for result
-      // For now, return the response and let user know it's processing
-      console.log('Video generation started. File ID:', data.file_id)
-      fs.writeFileSync(params.output, JSON.stringify(data, null, 2))
-      return { processing: true, file_id: data.file_id, saved: params.output }
-    } else if (
-      data.generatedSamples &&
-      Array.isArray(data.generatedSamples) &&
-      data.generatedSamples[0]?.video?.uri
-    ) {
-      // Vertex/Veo response: {generatedSamples: [{video: {uri: "https://..."}}]}
-      mediaUrl = data.generatedSamples[0].video.uri
-      mediaType = 'video'
-    } else if (data.videos && Array.isArray(data.videos) && data.videos[0]) {
-      // Vertex/Veo response: {videos: ["https://..."]}
-      mediaUrl = data.videos[0]
-      mediaType = 'video'
-    }
-
-    if (mediaUrl) {
-      // Download the actual media file from URL
-      const mediaResponse = await fetch(mediaUrl)
-      if (!mediaResponse.ok) {
-        throw new Error(
-          `Failed to download ${mediaType} from ${mediaUrl}: ${mediaResponse.status}`,
-        )
-      }
-      await saveBinaryResponse(mediaResponse, params.output)
-      return { saved: params.output, url: mediaUrl, type: mediaType }
-    }
-
-    fs.writeFileSync(params.output, JSON.stringify(data, null, 2))
-    return data
-  }
-
-  return apiHubPost('/run', request)
-}
-
-/**
- * Chat completion command
- * @param {object} params - Chat parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} [params.prompt] - Simple prompt (converted to messages)
- * @param {Array} [params.messages] - Full messages array
- * @param {string} [params.system] - System prompt
- * @param {boolean} [params.stream] - Enable streaming
- * @param {number} [params.maxTokens] - Max tokens
- * @param {number} [params.temperature] - Temperature
- * @returns {Promise<object|AsyncGenerator>} Chat response or stream
- */
-async function chat(params) {
-  let messages = params.messages
-  if (!messages && params.prompt) {
-    messages = [{ role: 'user', content: params.prompt }]
-  }
-  if (!messages) {
-    throw new Error('Either --prompt or --messages is required')
-  }
-
-  const inputs = { messages }
-  if (params.system) inputs.system = params.system
-  if (params.maxTokens) inputs.max_tokens = params.maxTokens
-  if (params.temperature !== undefined) inputs.temperature = params.temperature
-
-  return run({ model: params.model, inputs, stream: params.stream })
-}
-
-/**
- * Text-to-speech command
- * @param {object} params - TTS parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} params.text - Text to synthesize
- * @param {string} [params.voiceId] - Voice ID (provider-specific)
- * @param {string} params.output - Output audio file path
- * @returns {Promise<object>} TTS result
- */
-async function tts(params) {
-  if (!params.text) {
-    throw new Error('--text is required for TTS')
-  }
-  if (!params.output) {
-    throw new Error('--output is required for TTS')
-  }
-
-  const inputs = {}
-
-  // Provider-specific input mapping
-  const [vendor] = params.model.split('/')
-  if (vendor === 'elevenlabs') {
-    // ElevenLabs uses 'text' and requires voice_id - default to "Rachel"
-    inputs.text = params.text
-    inputs.voice_id = params.voiceId || 'EXAVITQu4vr4xnSDxMaL'
-  } else if (vendor === 'minimax') {
-    // MiniMax uses 'text' and 'voice_setting' object
-    inputs.text = params.text
-    inputs.voice_setting = {
-      voice_id: params.voiceId || 'male-qn-qingse',
-      speed: 1.0,
-      vol: 1.0,
-      pitch: 0,
-    }
-  } else if (vendor === 'openai') {
-    // OpenAI TTS uses 'input' and 'voice'
-    inputs.input = params.text
-    inputs.voice = params.voiceId || 'alloy'
-  } else if (vendor === 'replicate') {
-    // Replicate XTTS uses 'text' and requires 'speaker' (audio URL for voice cloning)
-    inputs.text = params.text
-    if (params.speaker) {
-      inputs.speaker = params.speaker
-    } else if (params.voiceId) {
-      inputs.speaker = params.voiceId
-    } else {
-      // Default speaker sample
-      inputs.speaker =
-        'https://replicate.delivery/pbxt/Jt79w0xsT64R1JsiJ0LQRL8UcWspg5J4RFrU6YwEKpOT1ukS/male.wav'
-    }
-  } else if (vendor === 'mm') {
-    // MM TTS (qwen3-tts-flash) uses 'text' and optional 'voice'
-    inputs.text = params.text
-    if (params.voiceId) {
-      inputs.voice = params.voiceId
-    }
-  } else {
-    // Default: use 'text'
-    inputs.text = params.text
-  }
-
-  return run({ model: params.model, inputs, output: params.output })
-}
-
-/**
- * Image generation command
- * @param {object} params - Image parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} params.prompt - Image generation prompt
- * @param {string} [params.size] - Image size (e.g., "1024x1024")
- * @param {string} [params.output] - Output file path
- * @returns {Promise<object>} Image generation result
- */
-async function image(params) {
-  if (!params.prompt) {
-    throw new Error('--prompt is required for image generation')
-  }
-
-  const [vendor] = params.model.split('/')
-  const inputs = {}
-
-  if (vendor === 'vertex') {
-    inputs.messages = [{ role: 'user', content: params.prompt }]
-  } else if (vendor === 'mm') {
-    // MM uses prompt and size in "1024*1536" format
-    inputs.prompt = params.prompt
-    if (params.size) {
-      // Convert "1024x1536" to "1024*1536" if needed
-      inputs.size = params.size.replace('x', '*')
-    }
-  } else {
-    inputs.prompt = params.prompt
-    if (params.size) inputs.size = params.size
-  }
-
-  return run({ model: params.model, inputs, output: params.output })
-}
-
-/**
- * Image upscale command (FAL creative-upscaler)
- * @param {object} params - Upscale parameters
- * @param {string} params.imageUrl - URL of image to upscale
- * @param {number} [params.scale] - Upscale factor: 2 or 4 (default: 2)
- * @param {string} [params.outputFormat] - "png" or "jpeg" (default: "png")
- * @param {string} [params.output] - Output file path
- * @returns {Promise<object>} Upscale result {image_url, images}
- */
-async function upscale(params) {
-  if (!params.imageUrl) {
-    throw new Error('--image-url is required for upscale')
-  }
-
-  const inputs = {
-    image_url: params.imageUrl,
-    scale: params.scale || 2,
-    output_format: params.outputFormat || 'png',
-  }
-
-  return run({ model: 'fal/upscale', inputs, output: params.output })
-}
-
-/**
- * Image-to-image transformation command (FAL FLUX dev)
- * @param {object} params - img2img parameters
- * @param {string} params.imageUrl - URL of source image
- * @param {string} params.prompt - Transformation description
- * @param {number} [params.strength] - Transform strength 0.0-1.0 (default: 0.75)
- * @param {string} [params.imageSize] - Size preset: square_hd, square, portrait_4_3, landscape_16_9, etc. (default: square_hd)
- * @param {string} [params.outputFormat] - "jpeg" or "png" (default: "jpeg")
- * @param {number} [params.numImages] - Number of images 1-4 (default: 1)
- * @param {string} [params.output] - Output file path
- * @returns {Promise<object>} img2img result {image_url, images}
- */
-async function img2img(params) {
-  if (!params.imageUrl) {
-    throw new Error('--image-url is required for img2img')
-  }
-  if (!params.prompt) {
-    throw new Error('--prompt is required for img2img')
-  }
-
-  const inputs = {
-    image_url: params.imageUrl,
-    prompt: params.prompt,
-    strength: params.strength || 0.75,
-    image_size: params.imageSize || 'square_hd',
-    output_format: params.outputFormat || 'jpeg',
-    num_images: params.numImages || 1,
-  }
-
-  return run({ model: 'fal/img2img', inputs, output: params.output })
-}
-
-/**
- * Web search command
- * @param {object} params - Search parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} params.query - Search query
- * @returns {Promise<object>} Search results
- */
-async function search(params) {
-  if (!params.query) {
-    throw new Error('--query is required for search')
-  }
-
-  const inputs = {}
-  const [vendor] = params.model.split('/')
-
-  // Provider-specific input mapping
-  if (vendor === 'scrapingdog') {
-    inputs.q = params.query
-  } else if (vendor === 'perplexity') {
-    // Perplexity uses chat-style interface
-    inputs.messages = [{ role: 'user', content: params.query }]
-  } else {
-    inputs.query = params.query
-  }
-
-  return run({ model: params.model, inputs })
-}
-
-/**
- * Web scraping command
- * @param {object} params - Scrape parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} [params.url] - Single URL to scrape
- * @param {string[]} [params.urls] - Multiple URLs to scrape
- * @returns {Promise<object>} Scrape results
- */
-async function scrape(params) {
-  if (!params.url && !params.urls) {
-    throw new Error('--url or --urls is required for scraping')
-  }
-
-  const inputs = {}
-  const [vendor] = params.model.split('/')
-
-  if (vendor === 'scrapingdog') {
-    inputs.url = params.url
-  } else if (vendor === 'firecrawl') {
-    if (params.urls) {
-      inputs.urls = params.urls
-    } else {
-      inputs.url = params.url
-    }
-  } else {
-    inputs.url = params.url
-  }
-
-  return run({ model: params.model, inputs })
-}
-
-/**
- * Linkup structured web search
- * @param {object} params - Linkup search parameters
- * @param {string} params.query - Search query
- * @param {string} [params.outputType] - "searchResults" | "sourcedAnswer" | "structured" (default: "searchResults")
- * @param {string} [params.depth] - "standard" | "deep" (default: "standard")
- * @param {string} [params.structuredOutputSchema] - JSON schema string for structured mode
- * @param {string[]} [params.includeDomains] - Only search these domains
- * @param {string[]} [params.excludeDomains] - Exclude these domains
- * @param {string} [params.fromDate] - Start date filter (YYYY-MM-DD)
- * @param {string} [params.toDate] - End date filter (YYYY-MM-DD)
- * @param {number} [params.maxResults] - Max results to return
- * @param {boolean} [params.includeImages] - Include images in results
- * @returns {Promise<object>} Search results in format matching outputType
- */
-async function linkupSearch(params) {
-  if (!params.query) {
-    throw new Error('--query is required for linkup-search')
-  }
-
-  const inputs = {
-    q: params.query,
-    depth: params.depth || 'standard',
-    outputType: params.outputType || 'searchResults',
-  }
-
-  if (params.structuredOutputSchema) inputs.structuredOutputSchema = params.structuredOutputSchema
-  if (params.includeDomains) inputs.includeDomains = params.includeDomains
-  if (params.excludeDomains) inputs.excludeDomains = params.excludeDomains
-  if (params.fromDate) inputs.fromDate = params.fromDate
-  if (params.toDate) inputs.toDate = params.toDate
-  if (params.maxResults) inputs.maxResults = params.maxResults
-  if (params.includeImages !== undefined) inputs.includeImages = params.includeImages
-
-  const model = params.depth === 'deep' ? 'linkup/search-deep' : 'linkup/search'
-  return run({ model, inputs })
-}
-
-/**
- * Linkup URL-to-markdown fetcher
- * @param {object} params - Linkup fetch parameters
- * @param {string} params.url - URL to fetch and convert to markdown
- * @param {boolean} [params.renderJs] - Render JavaScript before extracting (default: false)
- * @param {boolean} [params.includeImages] - Include images in output (default: false)
- * @param {boolean} [params.includeRawHtml] - Include raw HTML in response (default: false)
- * @returns {Promise<object>} { content, title, url }
- */
-async function linkupFetch(params) {
-  if (!params.url) {
-    throw new Error('--url is required for linkup-fetch')
-  }
-
-  const inputs = {
-    url: params.url,
-    renderJs: params.renderJs || false,
-    includeImages: params.includeImages || false,
-    includeRawHtml: params.includeRawHtml || false,
-  }
-
-  return run({ model: 'linkup/fetch', inputs })
-}
-
-/**
- * Video generation command
- * @param {object} params - Video parameters
- * @param {string} params.model - Model in "vendor/model" format
- * @param {string} params.prompt - Video generation prompt
- * @param {string} [params.output] - Output file path
- * @returns {Promise<object>} Video generation result
- */
-async function video(params) {
-  if (!params.prompt) {
-    throw new Error('--prompt is required for video generation')
-  }
-
-  const [vendor] = params.model.split('/')
-  const inputs = {}
-
-  if (vendor === 'vertex') {
-    // Vertex/Veo uses instances array format
-    inputs.instances = [{ prompt: params.prompt }]
-    inputs.parameters = {}
-  } else if (vendor === 'mm') {
-    // MM video models: t2v (text-to-video), i2v (image-to-video)
-    inputs.prompt = params.prompt
-    if (params.size) {
-      // Convert "1280x720" to "1280*720" if needed
-      inputs.size = params.size.replace('x', '*')
-    }
-    if (params.duration) {
-      inputs.duration = parseInt(params.duration)
-    }
-    if (params.image) {
-      // i2v mode: image-to-video
-      inputs.image_url = params.image
-    }
-  } else {
-    // MiniMax and others use 'prompt'
-    inputs.prompt = params.prompt
-  }
-
-  return run({ model: params.model, inputs, output: params.output })
-}
-
-/**
- * Multimodal understanding command (video/image/audio analysis)
- * @param {object} params - Multimodal parameters
- * @param {string} params.model - Model in "vendor/model" format (e.g., mm/qwen3-vl-plus)
- * @param {string} params.prompt - Text prompt/question about the media
- * @param {string} [params.video] - Video URL to analyze
- * @param {string} [params.image] - Image URL to analyze
- * @param {string} [params.audio] - Audio URL to analyze/transcribe
- * @returns {Promise<object>} Multimodal analysis result
- */
-async function multimodal(params) {
-  if (!params.prompt) {
-    throw new Error('--prompt is required for multimodal')
-  }
-  if (!params.video && !params.image && !params.audio) {
-    throw new Error('At least one of --video, --image, or --audio is required')
-  }
-
-  const [vendor] = params.model.split('/')
-  const inputs = {}
-
-  if (vendor === 'mm') {
-    // MM multimodal models use messages format
-    const content = []
-    if (params.video) {
-      content.push({ video: params.video })
-      if (params.fps) {
-        content[content.length - 1].fps = parseInt(params.fps)
-      }
-    }
-    if (params.image) {
-      content.push({ image: params.image })
-    }
-    if (params.audio) {
-      content.push({ audio: params.audio })
-    }
-    content.push({ text: params.prompt })
-
-    inputs.input = {
-      messages: [{ role: 'user', content }]
-    }
-  } else {
-    // Generic format
-    inputs.prompt = params.prompt
-    if (params.video) inputs.video_url = params.video
-    if (params.image) inputs.image_url = params.image
-    if (params.audio) inputs.audio_url = params.audio
-  }
-
-  return run({ model: params.model, inputs })
-}
-
-/**
- * Gamma presentation command
- * @param {object} params - Gamma parameters
- * @param {string} params.model - Model (gamma/generation)
- * @param {string} params.inputText - Presentation input text
- * @returns {Promise<object>} Gamma generation result
- */
-async function gamma(params) {
-  if (!params.inputText) {
-    throw new Error('--input-text is required for Gamma')
-  }
-
-  const inputs = {
-    inputText: params.inputText,
-    format: params.format || 'presentation',
-    textOptions: {
-      language: params.language || 'en'
-    }
-  }
-  return run({ model: params.model, inputs })
-}
-
-/**
- * Document processing command (Reducto)
- * @param {object} params - Document parameters
- * @param {string} params.model - Model in "reducto/model" format (parse, extract, split, edit)
- * @param {string} params.url - Document URL (PDF, DOCX, etc.)
- * @param {string} [params.schema] - JSON Schema string for extract (e.g. '{"type":"object","properties":{...}}')
- * @param {string} [params.splitDescription] - JSON array of {name, description} for split
- * @param {string} [params.instructions] - JSON string of edit instructions for edit
- * @param {string} [params.settings] - JSON string of additional settings
- * @param {string} [params.output] - Output file path to save results
- * @returns {Promise<object>} Document processing result
- */
-async function document(params) {
-  if (!params.url) {
-    throw new Error('--url is required (document URL)')
-  }
-
-  const inputs = { document_url: params.url }
-  const model = params.model.split('/')[1] // parse, extract, split, edit
-
-  if (model === 'extract' && params.schema) {
-    inputs.instructions = { schema: JSON.parse(params.schema) }
-  }
-  if (model === 'split' && params.splitDescription) {
-    inputs.split_description = JSON.parse(params.splitDescription)
-  }
-  if (model === 'edit' && params.instructions) {
-    inputs.edit_instructions = JSON.parse(params.instructions)
-  }
-  if (params.settings) {
-    inputs.settings = JSON.parse(params.settings)
-  }
-
-  return run({ model: params.model, inputs, output: params.output })
-}
-
-/**
- * Send SMS verification code (OTP) via Prelude
- * @param {object} params - Verify parameters
- * @param {string} params.phone - Phone number in E.164 format (e.g. "+1234567890")
- * @param {string} [params.ip] - User's IP address for anti-fraud signals
- * @param {string} [params.deviceId] - Device identifier for anti-fraud signals
- * @returns {Promise<object>} Verification result {id, status, method, channels}
- */
-async function smsVerify(params) {
-  if (!params.phone) {
-    throw new Error('--phone is required (E.164 format, e.g. +1234567890)')
-  }
-
-  const inputs = {
-    target: { type: 'phone_number', value: params.phone },
-  }
-  if (params.ip || params.deviceId) {
-    inputs.signals = {}
-    if (params.ip) inputs.signals.ip = params.ip
-    if (params.deviceId) inputs.signals.device_id = params.deviceId
-  }
-
-  return run({ model: 'prelude/verify-send', inputs })
-}
-
-/**
- * Check SMS verification code (OTP) via Prelude
- * @param {object} params - Check parameters
- * @param {string} params.phone - Phone number in E.164 format
- * @param {string} params.code - OTP code received via SMS
- * @returns {Promise<object>} Check result {id, status}
- */
-async function smsCheck(params) {
-  if (!params.phone) {
-    throw new Error('--phone is required (E.164 format, e.g. +1234567890)')
-  }
-  if (!params.code) {
-    throw new Error('--code is required (the OTP code received via SMS)')
-  }
-
-  const inputs = {
-    target: { type: 'phone_number', value: params.phone },
-    code: params.code,
-  }
-
-  return run({ model: 'prelude/verify-check', inputs })
-}
-
-/**
- * Send SMS notification via Prelude
- * @param {object} params - Notify parameters
- * @param {string} params.phone - Phone number in E.164 format
- * @param {string} params.templateId - Prelude template ID (configured in Prelude dashboard)
- * @param {object} [params.variables] - Template variables
- * @param {string} [params.from] - Sender number
- * @returns {Promise<object>} Send result
- */
-async function smsSend(params) {
-  if (!params.phone) {
-    throw new Error('--phone is required (E.164 format, e.g. +1234567890)')
-  }
-  if (!params.templateId) {
-    throw new Error('--template-id is required (configured in Prelude dashboard)')
-  }
-
-  const inputs = {
-    template_id: params.templateId,
-    to: params.phone,
-  }
-  if (params.variables) inputs.variables = params.variables
-  if (params.from) inputs.from = params.from
-
-  return run({ model: 'prelude/notify-send', inputs })
-}
-
-/**
- * List available models from API Hub
- * @param {object} [params] - List parameters
- * @param {string} [params.type] - Filter by category (chat, tts, image, video, scraping, etc.)
- * @param {string} [params.vendor] - Filter by vendor
- * @returns {Promise<object>} Models list
- */
-async function listModels(params = {}) {
-  const response = await apiHubGet('/v1/models')
-  let models = response.models || []
-
-  // Filter by category/type
-  if (params.type) {
-    const typeFilter = params.type.toLowerCase()
-    models = models.filter(m =>
-      m.category?.toLowerCase() === typeFilter ||
-      m.type?.toLowerCase() === typeFilter
-    )
-  }
-
-  // Filter by vendor
-  if (params.vendor) {
-    const vendorFilter = params.vendor.toLowerCase()
-    models = models.filter(m => m.vendor?.toLowerCase() === vendorFilter)
-  }
-
-  return { count: models.length, models }
-}
-
+const { fetchWithRetry } = require('./lib/fetch-retry')
+const { config } = require('./lib/client')
+
+// Commands
+const { run } = require('./commands/run')
+const { chat } = require('./commands/chat')
+const { tts } = require('./commands/tts')
+const { stt } = require('./commands/stt')
+const { image, upscale, img2img } = require('./commands/image')
+const { video, multimodal } = require('./commands/video')
+const { search, scrape, linkupSearch, linkupFetch } = require('./commands/search')
+const { sendEmail, sendBatchEmails } = require('./commands/email')
+const { smsVerify, smsCheck, smsSend } = require('./commands/sms')
+const { gamma, document } = require('./commands/document')
+const { listModels } = require('./commands/models')
 
 // CLI argument parsing
 function parseArgs(args) {
@@ -991,8 +90,9 @@ SkillBoss API Hub - Multi-Provider API Gateway
 Commands:
   list-models  List available models from API Hub
   run          Generic endpoint access (any model)
-  chat         Chat completions (bedrock, openai, anthropic, openrouter, vertex, minimax)
+  chat         Chat completions (bedrock, openai, anthropic, openrouter, vertex, minimax, huggingface)
   tts          Text-to-speech (elevenlabs, minimax, openai, mm/qwen3-tts-flash)
+  stt          Speech-to-text (openai/whisper-1, huggingface/openai/whisper-large-v3)
   image        Image generation (vertex/gemini, replicate/flux, mm/img)
   upscale      Image upscaling (fal/upscale)
   img2img      Image-to-image transformation (fal/img2img)
@@ -1024,10 +124,20 @@ Examples:
   # Chat
   node api-hub.js chat --model "bedrock/claude-4-sonnet" --prompt "Hello"
   node api-hub.js chat --model "openrouter/deepseek/deepseek-r1" --prompt "Hello" --stream
+  node api-hub.js chat --model "huggingface/meta-llama/Llama-3-8B-Instruct" --prompt "Hello"
+
+  # HuggingFace (any model from huggingface.co/models works)
+  node api-hub.js chat --model "huggingface/zai-org/GLM-5" --prompt "Hello" --stream
+  node api-hub.js run --model "huggingface/BAAI/bge-small-en-v1.5" --inputs '{"task":"embedding","input":"hello world"}'
+  node api-hub.js run --model "huggingface/stabilityai/stable-diffusion-xl-base-1.0" --inputs '{"task":"image","prompt":"a sunset"}' --output /tmp/image.png
 
   # TTS
   node api-hub.js tts --model "elevenlabs/eleven_multilingual_v2" --text "Hello" --output /tmp/audio.mp3
   node api-hub.js tts --model "mm/qwen3-tts-flash" --text "Hello" --output /tmp/audio.wav
+
+  # STT (Speech-to-Text)
+  node api-hub.js stt --file recording.mp3
+  node api-hub.js stt --file interview.wav --language en --output /tmp/transcript.txt
 
   # Multimodal (video/image understanding)
   node api-hub.js multimodal --model "mm/qwen3-vl-plus" --video "https://example.com/video.mp4" --prompt "What's happening in this video?"
@@ -1204,6 +314,25 @@ Examples:
           output: args.output,
         })
         console.log(`Audio saved to: ${args.output}`)
+        break
+      }
+
+      case 'stt': {
+        if (!args.file) {
+          console.error('Error: --file is required (local audio file path)')
+          process.exit(1)
+        }
+        result = await stt({
+          file: args.file,
+          model: args.model,
+          prompt: args.prompt,
+          language: args.language,
+          output: args.output,
+        })
+        console.log(result.text)
+        if (result.saved) {
+          console.log(`\nTranscript saved to: ${result.saved}`)
+        }
         break
       }
 
@@ -1586,6 +715,7 @@ module.exports = {
   run,
   chat,
   tts,
+  stt,
   image,
   upscale,
   img2img,
